Package scaffold (#6)
Set up the `bertron` package with [`uv`][0]

The package is organized like this:
```
.
├── pyproject.toml
├── src
│   └── bertron
│       ├── emsl
│       ├── ess_dive
│       ├── jgi
│       └── nmdc
└── uv.lock 
```

where `src/bertron` is the code for a package named `bertron` that
serves as a top-level entry point, and imports submodules from from
`src/bertron/{emsl,ess_dive,jgi,nmdc}` (one per data source) to do the
work.

I was trying out this project scaffold in
https://code.jgi.doe.gov/lirber/ber-branchwater-poc/, so the code that
actually exists is trying to search for `SRR` accessions in JGI (via
JDP) and `nmdc` (via `nmdc-api-utilities`). Likely a better initial
query is adapting the code for pulling lat/lon that we already have in
this repo.

[0]: https://docs.astral.sh/uv/