Implement a container-based development environment (#36)
On this branch, I implemented a FastAPI app having two endpoints and its
built-in Swagger UI documentation. I also containerized that FastAPI app
for deployment in both a development environment (e.g. laptop) and a
production environment (e.g. cloud hosting). The containerized
environment includes `uv` and all the Python dependencies specified in
`pyproject.toml`.

The reason I opted to put the `server.py` file directly in `src` is that
not all of the dependencies specified in `src/bertron` are accounted for
in `pyproject.toml` and I didn't want to resolve that issue as part of
this containerization effort (I'll create a new issue about that). I
opted to avoid the issue by not requiring Python to interpret the
contents of `src/bertron`.

---------

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>