Delete early data files/scripts and update map to fetch data from `data` repo (#44)
On this branch, I deleted the four, data source-specific directories
that we created early on in this repo's existence. We have since moved
them to a new repo, named `data`.

Deleting those broke the map web page, which was programmed to fetch the
data files from this repo. So, on this branch, I also updated the map
web page to fetch them from the `data` repo.

---------

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>