Implement GHA workflow that builds/pushes container image to GHCR (#50)
On this branch, I implemented a GitHub Actions workflow that—when
run—will build the `web` container image (its name will be
`bertron-api`) and publish it to the GitHub Container Registry. Once it
has been published there, NERSC's Spin platform will be able to pull it
from there and spin it up.

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>