Make API version number accessible via HTTP and display on swagger UI (#64)
On this branch, I update the GitHub Actions workflow that builds the
container image, so it gets the Git tag name (if the workflow was
triggered by a GitHub Release being published) or the Git commit hash
(otherwise), and overwrites the placeholder `"0.0.0"` version number in
`pyproject.toml`.

I also added a `/version` API endpoint that reads the value from
`pyproject.toml` and returns it to the HTTP client. This, effectively,
allows HTTP clients to determine which version of the API they are
accessing.

I also updated the existing `/health` API endpoint to use `true`/`false`
as its possible field values, instead of one field using `"ok"` and the
other field using `true`/`false`. In other words, I standardized the
values across fields. I also implemented a Pydantic model named
`HealthResponse`, which makes the "schema" of that response visible to
users of our Swagger UI page.

Finally, I added a `src/README.md` file that contains one-line
descriptions of the contents of the `src/` folder (except for one item,
which I'm punting on describing).

---------

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>