Display API version number at top of Swagger UI page (#66)
On this branch, I customized the Swagger UI header. Specifically, I made
it so the header shows the same package version number as the `/version`
endpoint returns, which comes from `pyproject.toml`. I also replaced the
default "FastAPI" title with "BERtron API". Finally, I added a
description that links to the GitHub repository and shows the schema
version.

---------

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>