Pin `bertron-schema` version to `v0.1.0-alpha.10` using Git tag (#91)
I'd like to switch to using the tagged version of the schema. 

As a one time fix I would also like to update the tag "v0.1.0-alpha.10"
on this repo to match this commit.

I realize that this is essentially re-tagging things, but it will allow
us to bring all the repos in sync

---------

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>