Bump `bertron-schema` tag to `v0.1.0-alpha.11` (#93)
On this branch, I updated the references to schema versions, to use a
newly-created tag: `v0.1.0-alpha.11`. Both the previous tag and this new
tag point to the same commit in the `bertron-schema` repository. The
changes in this PR are being made in pursuit of having all three repos
use the same tag for versions that are compatible with one another.

Co-authored-by: eecavanna <eecavanna@users.noreply.github.com>